new graph structure

+++ osm-tasks/src/config.rs
@@ -53,7 +53,7 @@ impl Config {
     pub fn force_rebuild_graph(&self) -> bool {
         self.force_rebuild_graph
     }
-    pub fn number_of_nodes(&self) -> u32 {
+    pub fn number_of_nodes(&self) -> usize {
         return 1000000;
     }
     pub fn build_graph_on_startup(&self) -> bool {
